requirements.txt
streamlit==1.38.0
pandas==2.2.2
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
pypdf==4.3.1
openpyxl==3.1.2
python-dotenv==1.0.1
dnspython==2.6.1

# Optional UI charts (lazy-import in app.py)
plotly==5.22.0

# Optional AI providers (lazy-import in ai_enrichment.py)
anthropic==0.34.2
groq==0.11.0
